Rework the code to allow path elements having the delimiter char inside
The behaviour how the path is evaluated changed!
Pass either an arrayref with your atomic path elements or pass a compact
string using the delimiter. Elements inside the arrayref are not split.
The private _build_path will now always return the path as a list and
never join the path into a string. Calls in scalar context cause a die.

If you want back the old behaviour, set RECURSEPATH => 1 in the constructor.
Be aware that this setting is NOT propagated to sub connectors when using
Connetor::Multi!