Fix bind issues in LDAP Auth Connector
To validate the password of a user in indirect mode the connector tries
to log in as the user with the supplied user password. This leads to the
situation that the global ldap handle is running with the users permissions
afterwards. This patch introduces a mandatory rebind after each try to
change the user.
This fixes openxpki/openxpki#870